# refactor(bb): Optimize interleaved polynomial batching in Gemini with parallel chunking