This PR optimizes the interleaved polynomial batching operation in Gemini's `compute_batched` method by applying parallel chunking. The optimization pre-computes scalar values and uses `parallel_for` with `ThreadChunk` to distribute the batching work across threads.

  Before the change:

  | CPUs | Time (ms) | Speedup | Efficiency |
  |------|-----------|---------|------------|
  | 1    | 799.74    | 1.00x   | 100.0%     |
  | 2    | 411.74    | 1.94x   | 97.0%      |
  | 4    | 217.65    | 3.67x   | 91.8%      |
  | 8    | 131.30    | 6.09x   | 76.1%      |
  | 16   | 114.82    | 6.97x   | 43.6%      |

  After the change:

  | CPUs | Time (ms) | Speedup | Efficiency |
  |------|-----------|---------|------------|
  | 1    | 799.41    | 1.00x   | 100.0%     |
  | 2    | 410.15    | 1.95x   | 97.5%      |
  | 4    | 216.32    | 3.70x   | 92.5%      |
  | 8    | 121.16    | 6.60x   | 82.5%      |
  | 16   | 111.17    | 7.19x   | 44.9%      |

  Performance Improvements:

  - 8 cores: 7.7% faster (131.30ms → 121.16ms)
  - 16 cores: 3.2% faster (114.82ms → 111.17ms)
  - Efficiency at 8 cores: Improved from 76.1% to 82.5%
  - Efficiency at 16 cores: Improved from 43.6% to 44.9%