@ludamad 2025-09-16
Nice speedup! I assume we have the same caveat here with 16 cores? it can be worth just getting those numbers from mainframe. I'll be adding a bench soon for them